requirements/base.txt
dj-database-url==0.4.1
Django==1.10.5
django-environ>=0.4.0
django-braces>=1.8.1
gunicorn==19.6.0
psycopg2==2.7.3.1
whitenoise==3.2
pytz==2017.3
tzlocal==1.5.1
py-moneyed==0.7.0
django-money==0.12.3
django-countries==5.0
django-bmemcached==0.2.3
python-binary-memcached==0.26.1
# django-heroku-memcacheify
python-dateutil==2.6.1
django-tables2==1.17.1
django_filter==1.1.0
tablib==0.12.1
django-bootstrap-breadcrumbs==0.8.2
pinax-stripe==4.0.0
django-crispy-forms==1.7.0
django-extensions==1.9.9
django-storages==1.6.5
easy-thumbnails==2.5
boto3==1.5.18
django-cors-headers==2.1.0
raven==6.5.0
django-amazon-ses==1.0.0
celery[redis]
django-fsm